Sync Computacional robot from Gabriel lunch-delivery-robot

+++ Computacional/robot/Marmitron/lib/ControlePID/ControlePID.h
@@ -14,7 +14,7 @@ class PIDController {
   public:
     PIDController(float kp, float ki, float kd);
     void setGanhos(float kp, float ki, float kd);
-    void setSentido(int sentido);
+    void setSentido(float sentido);
     int controle(float vel_ref, float vel_filtrada);
     // void setLimites(float min, float max);
     // float computar(float valor_alvo, float valor_atual);
